修复设置上方选择器颜色不正确的问题 调整弹幕速度 一些小细节的优化 弹幕不正确可以从播放页重新匹配

+++ DanDanPlayForiOS/NetManager/BaseNetManager.m
@@ -22,7 +22,7 @@ CG_INLINE NSError *humanReadableError(NSError *error) {
     
     NSHTTPURLResponse *response = error.userInfo[AFNetworkingOperationFailingURLResponseErrorKey];
     NSUInteger statusCode = response.statusCode;
-    NSMutableString *str = [[NSMutableString alloc] initWithFormat:@"%ld ", statusCode];
+    NSMutableString *str = [[NSMutableString alloc] initWithFormat:@"%lu ", (unsigned long)statusCode];
     if (statusCode == 400) {
         [str appendString:@"客户端发送的请求有错误"];
     }
